refactor: Sidebar
