+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 ## Under the Hood
 
 - Refactored the main editor component further
+- Refactored the Sidebar panels into their own respective components
 - Upgrade Electron to `v18.0.0`
 
 # 2.3.0
